// src/models/api/interrogation.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * queen-api
 * API for Queen/Stromae
 * OpenAPI spec version: 5.4.0
 */
import type { Data } from './data'
import type { InterrogationComment } from './interrogationComment'
import type { Personalization } from './personalization'
import type { StateData } from './stateData'

export interface Interrogation {
  id?: string
  questionnaireId?: string
  personalization?: Personalization
  data?: Data
  comment?: InterrogationComment
  stateData?: StateData
}